fix: plugin configuration error.
