--- pkg/filter/proxywasm/utils.go
@@ -1,33 +1,0 @@
-package proxywasm
-
-import (
-	"net/http"
-
-	"mosn.io/proxy-wasm-go-host/proxywasm/common"
-)
-
-// wrapper for http.Header, convert Header to api.HeaderMap.
-type myHeaderMap struct {
-	realMap http.Header
-}
-
-func (m *myHeaderMap) Get(key string) (string, bool) {
-	return m.realMap.Get(key), true
-}
-
-func (m *myHeaderMap) Set(key, value string) { panic("implemented") }
-
-func (m *myHeaderMap) Add(key, value string) { panic("implemented") }
-
-func (m *myHeaderMap) Del(key string) { panic("implemented") }
-
-func (m *myHeaderMap) Range(f func(key string, value string) bool) {
-	for k, _ := range m.realMap {
-		v := m.realMap.Get(k)
-		f(k, v)
-	}
-}
-
-func (m *myHeaderMap) Clone() common.HeaderMap { panic("implemented") }
-
-func (m *myHeaderMap) ByteSize() uint64 { panic("implemented") }
